Refactor the eBPF backend
